# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add an item to the cart (or increment qty if present)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartSummary"
  /cart/items/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartSummary"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove an item from the cart
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartSummary"
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "200":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
        "404":
          description: No account for this email
...
components:
  schemas:
...
    Product:
      type: object
      required:
        [
          id,
          vendorId,
          categoryId,
          slug,
          title,
          pricePesewas,
          stock,
          ratingAvgX100,
          ratingCount,
          isActive,
        ]
      properties:
        id: { type: integer }
        vendorId: { type: integer }
        categoryId: { type: integer }
        slug: { type: string }
        title: { type: string }
        brand: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        pricePesewas: { type: integer }
        compareAtPesewas: { type: ["integer", "null"] }
        stock: { type: integer }
        tag: { type: ["string", "null"] }
        ratingAvgX100: { type: integer }
        ratingCount: { type: integer }
        isActive: { type: boolean }
        imageUrl: { type: ["string", "null"] }
...
    CartItemInput:
      type: object
      required: [productId, qty]
      properties:
        productId: { type: integer }
        qty: { type: integer, minimum: 1 }
    CartItemUpdate:
      type: object
      required: [qty]
      properties:
        qty: { type: integer, minimum: 1 }
    CartLine:
      type: object
      required: [id, productId, qty, product]
      properties:
        id: { type: integer }
        productId: { type: integer }
        qty: { type: integer }
        product:
          $ref: "#/components/schemas/Product"
    CartSummary:
      type: object
      required: [id, items, subtotalPesewas]
      properties:
        id: { type: integer }
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartLine"
        subtotalPesewas: { type: integer }
    UserRoleAssignment:
      type: object
      required: [role, vendorId]
      properties:
        role:
          type: string
          enum: [buyer, vendor_owner, vendor_staff, admin, support_agent]
        vendorId: { type: ["integer", "null"] }
    AuthUser:
      type: object
      required: [id, email, phone, firstName, lastName, roles]
      properties:
        id: { type: integer }
        email: { type: string }
        phone: { type: ["string", "null"] }
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        roles:
          type: array
          items:
            $ref: "#/components/schemas/UserRoleAssignment"
...
    SignupInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
...